refactor : Subway 도메인별 패키지 분리
