+++ src/main/java/subway/domain/member/FavoriteRepository.java
@@ -1,6 +1,7 @@
-package subway.domain;
+package subway.domain.member;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import subway.domain.member.Favorite;
 
 public interface FavoriteRepository extends JpaRepository<Favorite, Long> {
