+++ src/main/java/subway/domain/member/MemberRepository.java
@@ -1,6 +1,7 @@
-package subway.domain;
+package subway.domain.member;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import subway.domain.member.Member;
 
 public interface MemberRepository extends JpaRepository<Member, Long> {
