+++ src/main/java/subway/domain/station/StationRepository.java
@@ -1,6 +1,7 @@
-package subway.domain;
+package subway.domain.station;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import subway.domain.station.Station;
 
 public interface StationRepository extends JpaRepository<Station, Long> {
