+++ src/test/java/subway/domain/FavoriteRepositoryTest.java
@@ -9,6 +9,10 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.boot.test.autoconfigure.orm.jpa.TestEntityManager;
+import subway.domain.member.Favorite;
+import subway.domain.member.Member;
+import subway.domain.member.FavoriteRepository;
+import subway.domain.member.MemberRepository;
 
 @DataJpaTest
 public class FavoriteRepositoryTest {
