+++ src/test/java/subway/domain/LineRepositoryTest.java
@@ -6,6 +6,10 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
+import subway.domain.line.Line;
+import subway.domain.line.LineRepository;
+import subway.domain.station.StationRepository;
+import subway.domain.station.Station;
 
 @DataJpaTest
 public class LineRepositoryTest {
