+++ src/test/java/subway/domain/MemberRepositoryTest.java
@@ -5,6 +5,10 @@ import static org.assertj.core.api.Assertions.assertThat;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
+import subway.domain.member.Favorite;
+import subway.domain.member.Member;
+import subway.domain.member.FavoriteRepository;
+import subway.domain.member.MemberRepository;
 
 @DataJpaTest
 public class MemberRepositoryTest {
