+++ src/test/java/subway/domain/StationRepositoryTest.java
@@ -6,9 +6,11 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
-import org.springframework.test.context.TestPropertySource;
+import subway.domain.line.Line;
+import subway.domain.line.LineRepository;
+import subway.domain.station.StationRepository;
+import subway.domain.station.Station;
 
 
 @DataJpaTest
